requirements.txt
Flask==2.2.2
gunicorn==20.0.4
Pillow~=9.4.0
tensorflow==2.10.0
numpy~=1.21.6
opencv-python~=4.5.3.56
requests~=2.28.2
scikit-learn